Update for schema v0.2
This removes the special case for filtering out .notes, which was
never specified in the v0.1 schema, but was considered "valid"
as a compromise with the existing data. In v0.2, most of the uses
for .notes are covered in new properties that *are* specified as
part of the schema: see opws/opws-schemata#1.

To validate documents that would have been previously considered
valid under this exemption, strict schema adherence must be disabled
with the `--loose` option. (Note that this will ignore any *other*
properties that are not defined by the schema as well.)

This closes issue #2.